First working attempt to use a full xpath-library for queries
A few things are quiet hard to do with apath queries because of differences. Not sure if we should switch to an actual library, since our old approach sort of worked.

XJPath has synchronized static locking in its context creation, and that's REALLY not necessary. It makes me wonder if this library is the right thing to use.